% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStreams.R
\name{CreateStreams}
\alias{CreateStreams}
\title{Create streams}
\usage{
CreateStreams(numWorkItems)
}
\arguments{
\item{x}{A vector indicating the number of work-items.}
}
\value{
A R-streams.
}
\description{
Create streams in R.
}
