% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFisher_sim.R
\name{fisher.sim}
\alias{fisher.sim}
\title{Fisher's exact test on GPU}
\usage{
fisher.sim(
  x,
  B,
  type = c("float", "double")[1 + gpuR::gpuInfo()$double_support],
  returnStatistics = FALSE,
  streams,
  workgroupSize,
  localSize = c(2, 2),
  verbose = FALSE
)
}
\description{
Fisher's exact test on GPU
}
