% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFisher_sim.R
\name{gpuFisher_sim}
\alias{gpuFisher_sim}
\title{Fisher's exact test on GPU}
\usage{
gpuFisher_sim(sr, sc, result, streams, workgroupSize, localSize = c(2,
  2), verbose = FALSE)
}
\description{
Fisher's exact test on GPU
}
