% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpuRnBackend}
\alias{gpuRnBackend}
\title{Random number generation}
\usage{
gpuRnBackend(x, streams, max_global_size, random_type)
}
\arguments{
\item{x}{output matrix}

\item{streams}{matrix of random number seeds}

\item{max_global_size}{vector of length 2, number of work items}

\item{random_type}{one of "uniform" or "normal"}
}
\description{
Fills a matrix with random numbers
}
